+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
-## 3.2.3.6
-- Fix NPE when trying to edit mods for an instance
-- Fix some content not being relocalized when language is changed
+## 3.2.3.7
+- Add in Alex default skins for even UUID's
+- Add in version_type to the launch arguments
+- Fix dependsGroup on extraArguments not working correctly
+- Stop showing error dialog when no skin is set by the user and is using default